refactor(region): extract MeetingsSyncService — #828 Step 2
Second bounded-context split. Extracts both syncMeetings (the
discover → extract_and_upsert → minutes_link phase pipeline) and
syncMeetingMinutes (the discover → ingest → summarize sub-pipeline)
into MeetingsSyncService. The minutes ingest stays inside the meetings
service because the operator-facing flow treats them as one lifecycle —
matches the phase-observability grouping (#825) verbatim.

region-sync.service.ts shrinks from 3,538 → 3,318 LOC (cumulative
3,929 → 3,318, −16% so far). MeetingsSyncService is 286 LOC.

No stage-taxonomy coupling here (meetings have no lifecycle stages),
so this extraction is cleaner than propositions — DbService + an
@Optional LegislativeActionLinkerService are the only deps. The
orchestrator passes its upsertByExternalId helper in as a callback;
behavior including the 'meetings:' cache prefix invalidation is
byte-identical.

All 2,105 backend tests pass.

Refs #828
